+++ src/db/spanner/pool.rs
@@ -37,7 +37,7 @@ pub struct SpannerDbPool {
     coll_cache: Arc<CollectionCache>,
 
     metrics: Metrics,
-    quota: u32,
+    quota: usize,
     quota_enabled: bool,
 }
 
@@ -58,7 +58,7 @@ impl SpannerDbPool {
             pool,
             coll_cache: Default::default(),
             metrics: metrics.clone(),
-            quota: settings.limits.max_quota_limit,
+            quota: settings.limits.max_quota_limit as usize,
             quota_enabled: settings.enable_quota,
         })
     }
